refactoring - move 'service' package out of 'jpa' parent. some formatting
