+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsinterventionsservice/events/NotifyListeners.kt
@@ -1,19 +1,19 @@
+// ktlint-disable filename
 package uk.gov.justice.digital.hmpps.hmppsinterventionsservice.events
 
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.context.ApplicationListener
 import org.springframework.stereotype.Component
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder
 import org.springframework.web.util.UriComponentsBuilder
-import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.jpa.service.NotifyService
+import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.service.NotifyService
 
 @Component
 class NotifyReferralEventListener(
   @Value("\${interventions.ui.uri}") private val interventionsUIURI: String,
   private val notifyService: NotifyService,
-): ApplicationListener<ReferralEvent> {
-  override fun onApplicationEvent(event: ReferralEvent){
-    when(event.type) {
+) : ApplicationListener<ReferralEvent> {
+  override fun onApplicationEvent(event: ReferralEvent) {
+    when (event.type) {
       ReferralEventType.SENT -> {
         // fixme: is it ok that this class implicitly knows the URL for the sent referral location?
         val location = UriComponentsBuilder.fromHttpUrl(interventionsUIURI)
